+++ src/XRP/wallet.ts
@@ -4,8 +4,12 @@ import * as rippleKeyPair from 'ripple-keypairs'
 
 import Utils from '../Common/utils'
 
+import XrpUtils from './xrp-utils'
+
 /**
  * An object which contains artifacts from generating a new wallet.
+ *
+ * @deprecated Use the WalletFactory class to vend wallets instead.
  */
 export interface WalletGenerationResult {
   /** The newly generated Wallet. */
@@ -53,6 +57,8 @@ class Wallet {
    *
    * Runtime environments that do not have secure random number generation should pass their own buffer of entropy.
    *
+   * @deprecated Please use methods on `WalletFactory` to generate wallets instead.
+   *
    * @param entropy - A optional hex string of entropy.
    * @param test - Whether the address is for use on a test network, defaults to `false`.
    * @returns Artifacts from the wallet generation.
@@ -83,6 +89,8 @@ class Wallet {
   /**
    * Generate a new hierarchical deterministic wallet from a mnemonic and derivation path.
    *
+   * @deprecated Please use methods on `WalletFactory` to generate wallets instead.
+   *
    * @param mnemonic - The given mnemonic for the wallet.
    * @param derivationPath - The given derivation path to use. If undefined, the default path is used.
    * @param test - Whether the address is for use on a test network, defaults to `false`.
@@ -109,6 +117,8 @@ class Wallet {
   /**
    * Generate a new hierarchical deterministic wallet from a seed and derivation path.
    *
+   * @deprecated Please use methods on `WalletFactory` to generate wallets instead.
+   *
    * @param seed - The given seed for the wallet.
    * @param derivationPath - The given derivation path to use. If undefined, the default path is used.
    * @param test - Whether the address is for use on a test network, defaults to `false`.
@@ -125,8 +135,8 @@ class Wallet {
       return undefined
     }
 
-    const publicKey = Wallet.hexFromBuffer(node.publicKey)
-    const privateKey = Wallet.hexFromBuffer(node.privateKey)
+    const publicKey = Utils.hexFromBuffer(node.publicKey)
+    const privateKey = Utils.hexFromBuffer(node.privateKey)
     return new Wallet(publicKey, `00${privateKey}`, test)
   }
 
@@ -152,24 +162,19 @@ class Wallet {
   }
 
   /**
-   * Converts a Buffer to an uppercase hexadecimal string.
-   *
-   * @param buffer - A Buffer to be converted to hexadecimal.
-   *
-   * @returns A hexadecimal string.
-   */
-  private static hexFromBuffer(buffer: Buffer): string {
-    return buffer.toString('hex').toUpperCase()
-  }
-
-  /**
    * Gets the x-address associated with a given wallet instance.
    *
    * @returns A string representing the x-address of the wallet.
+   *
+   * @throws An error if we are unable to derive an address.
    */
   public getAddress(): string {
     const classicAddress = rippleKeyPair.deriveAddress(this.publicKey)
-    const xAddress = Utils.encodeXAddress(classicAddress, undefined, this.test)
+    const xAddress = XrpUtils.encodeXAddress(
+      classicAddress,
+      undefined,
+      this.test,
+    )
     if (xAddress === undefined) {
       throw new Error('Unknown error deriving address')
     }